pyproject.toml
[tool.poetry]
name = "Solaris-Bot"
version = "1.0.0-beta.3"
description = "A Discord bot designed to make your server a safer and better place."

license = "GPLv3"
repository = "https://github.com/parafoxia/Solaris"
readme = "README.md"

authors = [
    "Ethan Henderson <parafoxia@carberra.xyz>"
]

[tool.poetry.dependencies]
python = "^3.8"
aiofiles = "^0.5.0"
aiosqlite = "^0.13.0"
apscheduler = "^3.6.3"
beautifultable = "^1.0.0"
"discord.py" = "1.3.4"
psutil = "^5.7.0"
pygount = "^1.2.3"
python-dotenv = "^0.14.0"
toml = "^0.10.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
mypy = "^0.782"

[tool.black]
line-length = 119
target-version = ["py38"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
